win: Use relative path for python in ninja files if possible
On Windows, `BuildSettings::python_path` returns the absolute path to
the script executable, even if it is within the root directory.  This
makes the ninja files non-portable because they include the absolute
path on the machine where `gn` was run.

This commit makes it such that if python is within the root directory,
we will use a relative path to it from the build directory when
generating the ninja rules.

This will allow us to setup siso rules to execute python scripts on
remote machines for Windows.

Note that we need a separate variable for this because the original
`python_path` is used for other purposes like executing scripts within
GN, so that needs to remain as an absolute path.

Change-Id: I1a1c2895efd868fb4cc6c9b8e2c61167bfe59211