Do not add .inputdeps paths to --ninja-outputs-file
This fixes a nasty bug where the content of --ninja-outputs-file
would mistakenly include the sources of copy() targets, as well as
intermediate `.inputdeps` Ninja targets created by the NinjaCopyTargetWriter.

This only happened when there were more than one generated source
listed in the copy() target's 'sources' argument.

The CL fixes the issue and adds a unit-test to ensure this does not
regress in the future.

Bug: 448860851
Change-Id: I195c2cc4b8b0ec5ef3902eaf6d831b220fd3c52c